// projects/library/date/src/lib/tsc/enum/date.enum.ts
export enum DateInputTypeEnum {
    DAY_INPUT = 'DAY_INPUT',
    MONTH_INPUT = 'MONTH_INPUT',
    YEAR_INPUT = 'YEAR_INPUT',
    CUSTOM_INPUT = 'CUSTOM_INPUT'
}

export enum DateTypeEnum {
    INCLUSIVE = 'INCLUSIVE',
    EXCLUSIVE = 'EXCLUSIVE',
}

export enum DateFormatEnum {
    DATE_TIME = 'DD-MMM-YYYY HH:mm:ss',
    DD_MMM_YYYY = 'DD-MMM-YYYY',
    YYYY_MM_DD = 'YYYY-MM-DD',
    MMM_YYYY = 'MMM YYYY'
}

export enum DateTimeIntervalEnum {
    CURRENT_TIME = 'CURRENT_TIME',
    CURRENT_HOUR = 'CURRENT_HOUR',
    LAST_24_HRS = 'LAST_24_HRS',
    YESTERDAY = 'YESTERDAY',
    CURRENT_DAY = 'CURRENT_DAY',
    NEXT_DAY = 'NEXT_DAY',
    LAST_7_DAYS = 'LAST_7_DAYS',
    CURRENT_WEEK = 'CURRENT_WEEK',
    NEXT_WEEK = 'NEXT_WEEK',
    LAST_30_DAYS = 'LAST_30_DAYS',
    CURRENT_MONTH = 'CURRENT_MONTH',
    NEXT_MONTH = 'NEXT_MONTH',
    LAST_12th_MONTH = 'LAST_12th_MONTH',
    CURRENT_YEAR = 'CURRENT_YEAR',
    LAST_12_MONTHS = 'LAST_12_MONTHS',
}